Make the search bar "functional", for some definition of "functional"
It's more than a little janky, in part because I had no time to properly
modify the existing JavaScript code to fit my needs. Instead, I deleted
it and kept only what's necessary to create a minimal version of the
functionality that still technically checks all the boxes we really need
check.