Put an upper limit on some loops
Before 707846a657a96ec7f729395ec110d0a70ca0b505, one of these loops
would cause an indefinite hang in the case where a newly registered user
tries to log in. This happens because most of the foreign keys would
point to User and not UserAccount. This is particularly bad for the
Session table, since we already interpret `IntegrityError`s as a token
collision.